pass first stage

+++ src/Server.cpp
@@ -126,7 +126,7 @@ int processVarInt(std::vector<char> &database_file ,unsigned short rowAddress){
 
 int getRowData(std::vector<char> &database_file , unsigned long rowAddress,std::vector<std::vector<std::string>> &tableData){
 
-    int next = rowAddress;
+    unsigned long next = rowAddress;
     std::vector<int> header; 
     
     next = processVarInt(database_file,rowAddress);
